Refactoring stuff around a bit. Sorry Kaizen :)
