+++ packages/jsii-rosetta/lib/jsii/assemblies.ts
@@ -13,6 +13,7 @@ import {
   ApiLocation,
 } from '../snippet';
 import { enforcesStrictMode } from '../strict';
+import { mkDict } from '../util';
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires, @typescript-eslint/no-require-imports
 const sortJson = require('sort-json');
@@ -186,4 +187,73 @@ function _fingerprint(assembly: spec.Assembly): spec.Assembly {
   assembly = sortJson(assembly);
   const fingerprint = crypto.createHash('sha256').update(JSON.stringify(assembly)).digest('base64');
   return { ...assembly, fingerprint };
+}
+
+export interface TypeLookupAssembly {
+  readonly assembly: spec.Assembly;
+  readonly assemblyFile: string;
+  readonly symbolIdMap: Record<string, string>;
+}
+
+const MAX_ASM_CACHE = 3;
+const ASM_CACHE: TypeLookupAssembly[] = [];
+
+/**
+ * Recursively searches for a .jsii file in the directory.
+ * When file is found, checks cache to see if we already
+ * stored the assembly in memory. If not, we synchronously
+ * load the assembly into memory.
+ */
+export function findTypeLookupAssembly(directory: string): TypeLookupAssembly | undefined {
+  const pjLocation = findPackageJsonLocation(path.resolve(directory));
+  if (!pjLocation) {
+    return undefined;
+  }
+
+  const assemblyFile = path.join(path.dirname(pjLocation), '.jsii');
+
+  const fromCache = ASM_CACHE.find((c) => c.assemblyFile === assemblyFile);
+  if (fromCache) {
+    return fromCache;
+  }
+
+  if (!fs.existsSync(assemblyFile)) {
+    return undefined;
+  }
+
+  const loaded = loadLookupAssembly(assemblyFile);
+  while (ASM_CACHE.length >= MAX_ASM_CACHE) {
+    ASM_CACHE.pop();
+  }
+  ASM_CACHE.unshift(loaded);
+  return loaded;
+}
+
+function loadLookupAssembly(assemblyFile: string): TypeLookupAssembly {
+  const assembly: spec.Assembly = fs.readJSONSync(assemblyFile, { encoding: 'utf-8' });
+  const symbolIdMap = mkDict(
+    Object.values(assembly.types ?? {}).map((type) => [type.symbolId ?? '', type.fqn] as const),
+  );
+
+  return {
+    assembly,
+    assemblyFile,
+    symbolIdMap,
+  };
+}
+
+function findPackageJsonLocation(currentPath: string): string | undefined {
+  // eslint-disable-next-line no-constant-condition
+  while (true) {
+    const candidate = path.join(currentPath, 'package.json');
+    if (fs.existsSync(candidate)) {
+      return candidate;
+    }
+
+    const parentPath = path.resolve(currentPath, '..');
+    if (parentPath === currentPath) {
+      return undefined;
+    }
+    currentPath = parentPath;
+  }
 }
