+++ packages/jsii-rosetta/package.json
@@ -32,7 +32,7 @@
   },
   "dependencies": {
     "@jsii/check-node": "0.0.0",
-    "@jsii/spec": "^0.0.0",
+    "@jsii/spec": "0.0.0",
     "commonmark": "^0.30.0",
     "fs-extra": "^9.1.0",
     "typescript": "~3.9.10",
@@ -40,7 +40,7 @@
     "@xmldom/xmldom": "^0.7.5",
     "workerpool": "^6.1.5",
     "yargs": "^16.2.0",
-    "jsii": "^.0.0"
+    "jsii": "0.0.0"
   },
   "license": "Apache-2.0",
   "author": {
