+++ packages/jsii/lib/index.ts
@@ -1,2 +1,3 @@
 export * from './jsii-diagnostic';
+export * from './symbol-id';
 export * from './helpers';
